# api/openapi.yaml
openapi: 3.0.3
info:
  title: Мини анкета
  description: API для мини-анкеты
  version: 1.0.0

servers:
  - url: http://localhost:8080
    description: Local development server

paths:
  /questions:
    get:
      operationId: getQuestions
      summary: Получить список вопросов анкеты
      tags: [Questions]
      responses:
        "200":
          description: Список вопросов
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Question"

  /answers:
    post:
      operationId: submitAnswers
      summary: Принять и сохранить ответы пользователя
      tags: [Answers]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/SubmitAnswersRequest"
      responses:
        "201":
          description: Ответы сохранены
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AnswerRecord"
        "400":
          $ref: "#/components/responses/BadRequest"

components:
  schemas:
    QuestionType:
      type: string
      enum: [text, single_choice, multiple_choice]
      description: Тип вопроса

    Question:
      type: object
      required: [id, text, type]
      properties:
        id:
          type: string
          format: uuid
          example: 550e8400-e29b-41d4-a716-446655440000
        text:
          type: string
          example: Какой ваш любимый цвет?
        type:
          $ref: "#/components/schemas/QuestionType"
        options:
          type: array
          items:
            type: string
          description: Варианты ответа для choice-типов
          example: ["Красный", "Синий", "Зелёный"]
        required:
          type: boolean
          default: false

    Answer:
      type: object
      required: [question_id, value]
      properties:
        question_id:
          type: string
          format: uuid
          description: ID вопроса
        value:
          oneOf:
            - type: string
            - type: array
              items:
                type: string
          description: |
            Для text: строка.
            Для single_choice: строка с выбранным вариантом.
            Для multiple_choice: массив строк.

    SubmitAnswersRequest:
      type: object
      required: [answers]
      properties:
        respondent:
          type: string
          maxLength: 200
          description: Имя респондента (опционально)
        answers:
          type: array
          minItems: 1
          items:
            $ref: "#/components/schemas/Answer"

    AnswerRecord:
      type: object
      required: [id, answers, submitted_at]
      properties:
        id:
          type: string
          format: uuid
        respondent:
          type: string
          nullable: true
        answers:
          type: array
          items:
            $ref: "#/components/schemas/Answer"
        submitted_at:
          type: string
          format: date-time

    Error:
      type: object
      required: [error]
      properties:
        error:
          type: string
          example: Bad request
        details:
          type: string

  responses:
    BadRequest:
      description: Неверный запрос
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"